requirements.txt
# Core web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# PDF generation
reportlab>=4.1.0

# Google Drive
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# HTTP client (async)
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0  # also used for TestClient

# Dev / optional
python-multipart>=0.0.9  # for FastAPI form data if needed
